Organize prob/, make problems request B fields
The idea is to have every problem capable of either B field transport,
but this requires defining a function for A or B rather than setting it,
since you don't know where it lives.
This solution is not ideal, but I think it'll work okay
