+++ kharma/prob/mhdmodes.hpp
@@ -68,51 +68,38 @@ TaskStatus InitializeMHDModes(std::shared_ptr<MeshBlockData<Real>>& rc, Paramete
     const auto& G = pmb->coords;
 
     const int nmode = pin->GetOrAddInteger("mhdmodes", "nmode", 1);
-    const int dir = pin->GetOrAddInteger("mhdmodes", "dir", 0);
     const bool one_period = pin->GetOrAddBoolean("mhdmodes", "one_period", true);
 
-    // if (pin->GetInteger("parthenon/mesh", "nx1")) {
-    //     dir = 3;
-    // }
-
-    // START POSSIBLE ARGS: take all these as parameters in pin?
     // Mean state
-    Real rho0 = 1.;
-    Real u0 = 1.;
-    Real u10 = 0.;
-    Real u20 = 0.;
-    Real u30 = 0.;
+    const Real rho0 = pin->GetOrAddReal("mhdmodes", "rho0", 1.);
+    const Real u0 = pin->GetOrAddReal("mhdmodes", "u0", 1.);
+    const Real u10 = pin->GetOrAddReal("mhdmodes", "u10", 0.);
+    const Real u20 = pin->GetOrAddReal("mhdmodes", "u20", 0.);
+    const Real u30 = pin->GetOrAddReal("mhdmodes", "u30", 0.);
 
-    // Wavevector
-    Real k1 = 2. * M_PI;
-    Real k2 = 2. * M_PI;
-    Real k3 = 2. * M_PI;
-    // "Faux-2D" plane orientation
-    // Set to 0 for "full" 3D wave
-    if (dir == 1)
-        k1 = 0;
-    if (dir == 2)
-        k2 = 0;
-    if (dir == 3)
-        k3 = 0;
+    // Wave parameters
+    // dir sets "Faux-2D" plane orientation, good for asymmetry bugs
+    // Set to 0 for "full" 3D wave.
+    const int dir = pin->GetOrAddInteger("mhdmodes", "dir", 0);
+    const Real amp = pin->GetOrAddReal("mhdmodes", "amp", 1.e-4);
 
-    Real amp = 1.e-4;
-    // END POSSIBLE ARGS
-
-    // B is set later, see below
-    Real B10 = 0.;
-    Real B20 = 0.;
-    Real B30 = 0.;
+    // Note the modes below don't work right if you manually set these
+    // TODO generate modes on the fly for any k values
+    const Real k1 = pin->GetOrAddReal("mhdmodes", "k1", (dir == 1) ? 0. : 2. * M_PI);
+    const Real k2 = pin->GetOrAddReal("mhdmodes", "k2", (dir == 2) ? 0. : 2. * M_PI);
+    const Real k3 = pin->GetOrAddReal("mhdmodes", "k3", (dir == 3) ? 0. : 2. * M_PI);
+    // Likewise
+    const Real B10 = pin->GetOrAddReal("mhdmodes", "B10", (dir == 0 || dir == 3) ? 1.0 : 0. );
+    const Real B20 = pin->GetOrAddReal("mhdmodes", "B20", (dir == 1) ? 1.0 : 0. );
+    const Real B30 = pin->GetOrAddReal("mhdmodes", "B30", (dir == 2) ? 1.0 : 0. );
 
     std::complex<Real> omega;
     Real drho = 0, du = 0;
     Real du1 = 0, du2 = 0, du3 = 0;
     Real dB1 = 0, dB2 = 0, dB3 = 0;
-
     // Eigenmode definitions
     if (dir == 0) {
         // 3D (1,1,1) wave
-        B10 = 1.;
         if (nmode == 0) { // Entropy
             drho = 1.;
         } else if (nmode == 1) { // Slow
@@ -146,15 +133,6 @@ TaskStatus InitializeMHDModes(std::shared_ptr<MeshBlockData<Real>>& rc, Paramete
     else
     {
         // 2D (1,1,0), (1,0,1), (0,1,1) wave
-        // Constant field direction
-        if (dir == 1) {
-            B20 = 1.;
-        } else if (dir == 2) {
-            B30 = 1.;
-        } else if (dir == 3) {
-            B10 = 1.;
-        }
-
         if (nmode == 0) { // Entropy
             drho = 1.;
         } else if (nmode == 1) { // Slow
@@ -212,6 +190,31 @@ TaskStatus InitializeMHDModes(std::shared_ptr<MeshBlockData<Real>>& rc, Paramete
         }
     }
 
+    // Record the parameters
+    // This might be useful to read when checking, too...
+    // TODO 
+    pin->SetReal("mhdmodes", "omega_real", omega.real());
+    pin->SetReal("mhdmodes", "omega_imag", omega.imag());
+
+    pin->SetReal("mhdmodes", "drho", drho);
+    pin->SetReal("mhdmodes", "du", du);
+    pin->SetReal("mhdmodes", "du1", du1);
+    pin->SetReal("mhdmodes", "du2", du2);
+    pin->SetReal("mhdmodes", "du3", du3);
+    pin->SetReal("mhdmodes", "dB1", dB1);
+    pin->SetReal("mhdmodes", "dB2", dB2);
+    pin->SetReal("mhdmodes", "dB3", dB3);
+
+    // Set B field parameters for our mode
+    pin->GetOrAddString("b_field", "type", "wave");
+    pin->GetOrAddReal("b_field", "b10", B10);
+    pin->GetOrAddReal("b_field", "b20", B20);
+    pin->GetOrAddReal("b_field", "b30", B30);
+    pin->GetOrAddReal("b_field", "amp_B1", amp*dB1);
+    pin->GetOrAddReal("b_field", "amp_B2", amp*dB2);
+    pin->GetOrAddReal("b_field", "amp_B3", amp*dB3);
+    pin->GetOrAddReal("b_field", "phase", 0.);
+
     IndexDomain domain = IndexDomain::interior;
     IndexRange ib = pmb->cellbounds.GetBoundsI(domain);
     IndexRange jb = pmb->cellbounds.GetBoundsJ(domain);
@@ -220,16 +223,12 @@ TaskStatus InitializeMHDModes(std::shared_ptr<MeshBlockData<Real>>& rc, Paramete
         KOKKOS_LAMBDA (const int &k, const int &j, const int &i) {
             Real X[GR_DIM];
             G.coord_embed(k, j, i, Loci::center, X);
-
-            Real mode = amp * cos(k1 * X[1] + k2 * X[2] + k3 * X[3]);
+            Real mode = amp * m::cos(k1 * X[1] + k2 * X[2] + k3 * X[3]);
             rho(k, j, i) = rho0 + drho * mode;
             u(k, j, i) = u0 + du * mode;
             uvec(0, k, j, i) = u10 + du1 * mode;
             uvec(1, k, j, i) = u20 + du2 * mode;
             uvec(2, k, j, i) = u30 + du3 * mode;
-            B_P(0, k, j, i) = B10 + dB1 * mode;
-            B_P(1, k, j, i) = B20 + dB2 * mode;
-            B_P(2, k, j, i) = B30 + dB3 * mode;
         }
     );
