+++ kharma/prob/seed_B.cpp
@@ -33,10 +33,12 @@
  */
 #include "seed_B.hpp"
 
-#include "seed_B_impl.hpp"
+#include "b_ct.hpp"
+#include "b_flux_ct.hpp"
 
 #include "boundaries.hpp"
 #include "coordinate_utils.hpp"
+#include "domain.hpp"
 #include "fm_torus.hpp"
 #include "grmhd_functions.hpp"
 
@@ -74,6 +76,268 @@ Real MinBeta(MeshData<Real> *md)
     return Reductions::DomainReduction<Reductions::Var::beta, Real>(md, UserHistoryOperation::min);
 }
 
+
+template <BSeedType Seed>
+TaskStatus SeedBFieldType(MeshBlockData<Real> *rc, ParameterInput *pin, IndexDomain domain = IndexDomain::entire)
+{
+    auto pmb = rc->GetBlockPointer();
+    auto pkgs = pmb->packages.AllPackages();
+
+    // Fields
+    GridScalar rho = rc->Get("prims.rho").data;
+    const auto &G = pmb->coords;
+
+    // Parameters
+    std::string b_field_type = pin->GetString("b_field", "type");
+    auto prob = pin->GetString("parthenon/job", "problem_id");
+    bool is_torus = (prob == "torus");
+
+    // Indices
+    // TODO handle filling faces with domain < entire more gracefully
+    IndexRange3 b = KDomain::GetRange(rc, domain);
+    int ndim = pmb->pmy_mesh->ndim;
+
+    // Shortcut to field values for easy fields
+    if constexpr (Seed == BSeedType::constant ||
+                  Seed == BSeedType::monopole ||
+                  Seed == BSeedType::monopole_cube ||
+                  Seed == BSeedType::orszag_tang)
+    {
+        // All custom B fields should set what they need of these
+        const Real b10 = pin->GetOrAddReal("b_field", "B10", 0.);
+        const Real b20 = pin->GetOrAddReal("b_field", "B20", 0.);
+        const Real b30 = pin->GetOrAddReal("b_field", "B30", 0.);
+        const Real k1 = pin->GetOrAddReal("b_field", "k1", 0.);
+        const Real k2 = pin->GetOrAddReal("b_field", "k2", 0.);
+        const Real k3 = pin->GetOrAddReal("b_field", "k3", 0.);
+        const Real phase = pin->GetOrAddReal("b_field", "phase", 0.);
+        const Real amp_B1 = pin->GetOrAddReal("b_field", "amp_B1", 0.);
+        const Real amp_B2 = pin->GetOrAddReal("b_field", "amp_B2", 0.);
+        const Real amp_B3 = pin->GetOrAddReal("b_field", "amp_B3", 0.);
+        const Real amp2_B1 = pin->GetOrAddReal("b_field", "amp2_B1", 0.);
+        const Real amp2_B2 = pin->GetOrAddReal("b_field", "amp2_B2", 0.);
+        const Real amp2_B3 = pin->GetOrAddReal("b_field", "amp2_B3", 0.);
+
+        if (pkgs.count("B_CT")) {
+            auto B_Uf = rc->PackVariables(std::vector<std::string>{"cons.fB"});
+            // Fill at 3 different locations
+            pmb->par_for(
+                "B_field_B", b.ks, b.ke, b.js, b.je, b.is, b.ie,
+                KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                    GReal Xembed[GR_DIM];
+                    double null1, null2;
+                    double B_Pf1, B_Pf2, B_Pf3;
+                    G.coord_embed(k, j, i, Loci::face1, Xembed);
+                    GReal gdet = G.gdet(Loci::face1, j, i);
+                    B_Pf1 = b10;
+                    seed_b<Seed>(Xembed, gdet, k1, k2, k3, phase,
+                                 amp_B1, amp_B2, amp_B3,
+                                 amp2_B1, amp2_B2, amp2_B3,
+                                 B_Pf1, null1, null2);
+                    B_Uf(F1, 0, k, j, i) = B_Pf1 * gdet;
+
+                    G.coord_embed(k, j, i, Loci::face2, Xembed);
+                    gdet = G.gdet(Loci::face2, j, i);
+                    B_Pf2 = b20;
+                    seed_b<Seed>(Xembed, gdet, k1, k2, k3, phase,
+                                 amp_B1, amp_B2, amp_B3,
+                                 amp2_B1, amp2_B2, amp2_B3,
+                                 null1, B_Pf2, null2);
+                    B_Uf(F2, 0, k, j, i) = B_Pf2;
+
+                    G.coord_embed(k, j, i, Loci::face3, Xembed);
+                    gdet = G.gdet(Loci::face3, j, i);
+                    B_Pf3 = b30;
+                    seed_b<Seed>(Xembed, gdet, k1, k2, k3, phase,
+                                 amp_B1, amp_B2, amp_B3,
+                                 amp2_B1, amp2_B2, amp2_B3,
+                                 null1, null2, B_Pf3);
+                    B_Uf(F3, 0, k, j, i) = B_Pf3 * gdet;
+                }
+            );
+            // Update primitive variables
+            B_CT::BlockUtoP(rc, domain);
+        } else if (pkgs.count("B_FluxCT")) {
+            GridVector B_P = rc->Get("prims.B").data;
+            pmb->par_for(
+                "B_field_B", b.ks, b.ke, b.js, b.je, b.is, b.ie,
+                KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                    GReal Xembed[GR_DIM];
+                    G.coord_embed(k, j, i, Loci::center, Xembed);
+                    const GReal gdet = G.gdet(Loci::center, j, i);
+                    seed_b<Seed>(Xembed, gdet, k1, k2, k3, phase,
+                                 amp_B1, amp_B2, amp_B3,
+                                 amp2_B1, amp2_B2, amp2_B3,
+                                 B_P(V1, k, j, i),
+                                 B_P(V2, k, j, i),
+                                 B_P(V3, k, j, i));
+                }
+            );
+            // We still need to update conserved flux values, but then we're done
+            B_FluxCT::BlockPtoU(rc, domain);
+        }
+        return TaskStatus::complete;
+    } else { // Seed with vector potential A otherwise
+        // Require and load what we need if necessary
+        Real A0 = pin->GetOrAddReal("b_field", "A0", 0.);
+        Real min_A = pin->GetOrAddReal("b_field", "min_A", 0.2);
+        // Init-specific loads
+        Real a, rin, rmax, gam, kappa, rho_norm, arg1;
+        Real tilt = 0; // Needs to be initialized
+        switch (Seed) {
+        case BSeedType::sane:
+        case BSeedType::mad:
+        case BSeedType::mad_quadrupole:
+        case BSeedType::r3s3:
+        case BSeedType::r5s5:
+        case BSeedType::gaussian:
+            // Torus parameters
+            rin = pin->GetReal("torus", "rin");
+            rmax = pin->GetReal("torus", "rmax");
+            kappa = pin->GetReal("torus", "kappa");
+            tilt = pin->GetReal("torus", "tilt") / 180. * M_PI;
+            // Other things we need only for torus evaluation
+            gam = pmb->packages.Get("GRMHD")->Param<Real>("gamma");
+            rho_norm = pmb->packages.Get("GRMHD")->Param<Real>("rho_norm");
+            a = G.coords.get_a();
+            break;
+        case BSeedType::orszag_tang_a:
+            A0 = pin->GetReal("orszag_tang", "tscale");
+            arg1 = pin->GetReal("orszag_tang", "phase");
+            break;
+        default:
+            break;
+        }
+
+        // For all other fields...
+        // Find the magnetic vector potential.  In X3 symmetry only A_phi is non-zero,
+        // But for tilted conditions we must keep track of all components
+        IndexSize3 sz = KDomain::GetBlockSize(rc);
+        ParArrayND<double> A("A", NVEC, sz.n3, sz.n2, sz.n1);
+        pmb->par_for(
+            "B_field_A", b.ks, b.ke, b.js, b.je, b.is, b.ie,
+            KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                GReal Xnative[GR_DIM];
+                GReal Xembed[GR_DIM], Xmidplane[GR_DIM];
+                G.coord(k, j, i, Loci::corner, Xnative);
+                G.coord_embed(k, j, i, Loci::corner, Xembed);
+                // What are our corresponding "midplane" values for evaluating the function?
+                rotate_polar(Xembed, tilt, Xmidplane);
+                const GReal r = Xmidplane[1], th = Xmidplane[2];
+
+                // In case we need zone sizes
+                const GReal dxc[GR_DIM] = {0., G.Dxc<1>(i), G.Dxc<2>(j), G.Dxc<3>(k)};
+
+                // This is written under the assumption re-computed rho is more accurate than a bunch
+                // of averaging in a meaningful way.  Just use the average if not.
+                Real rho_av;
+                if (is_torus) {
+                    // Find rho at corner directly for torii
+                    rho_av = fm_torus_rho(a, rin, rmax, gam, kappa, r, th) / rho_norm;
+                } else {
+                    // Use averages for anything else
+                    // This loop runs over every corner. Centers do not exist before the first
+                    // or after the last, so use the last (ghost) zones available.
+                    const int ii = clip((uint)i, b.is + 1, b.ie);
+                    const int jj = clip((uint)j, b.js + 1, b.je);
+                    const int kk = clip((uint)k, b.ks + 1, b.ke);
+                    if (ndim > 2)
+                    {
+                        rho_av = (rho(kk, jj, ii) + rho(kk, jj, ii - 1) +
+                                rho(kk, jj - 1, ii) + rho(kk, jj - 1, ii - 1) +
+                                rho(kk - 1, jj, ii) + rho(kk - 1, jj, ii - 1) +
+                                rho(kk - 1, jj - 1, ii) + rho(kk - 1, jj - 1, ii - 1)) /
+                                8;
+                    }
+                    else
+                    {
+                        rho_av = (rho(kk, jj, ii) + rho(kk, jj, ii - 1) +
+                                rho(kk, jj - 1, ii) + rho(kk, jj - 1, ii - 1)) /
+                                4;
+                    }
+                }
+
+                Real Aphi = seed_a<Seed>(Xmidplane, dxc, rho_av, rin, min_A, A0, arg1);
+
+                if (tilt != 0.0) {
+                    // This is *covariant* A_mu of an untilted disk
+                    const double A_untilt_lower[GR_DIM] = {0., 0., 0., Aphi};
+                    // Raise to contravariant vector, since rotate_polar_vec will need that.
+                    // Note we have to do this in the midplane!
+                    // The coord_to_native calculation involves an iterative solve for MKS/FMKS
+                    GReal Xnative_midplane[GR_DIM] = {0}, gcon_midplane[GR_DIM][GR_DIM] = {0};
+                    G.coords.coord_to_native(Xmidplane, Xnative_midplane);
+                    G.coords.gcon_native(Xnative_midplane, gcon_midplane);
+                    double A_untilt[GR_DIM] = {0};
+                    DLOOP2 A_untilt[mu] += gcon_midplane[mu][nu] * A_untilt_lower[nu];
+
+                    // Then rotate
+                    double A_tilt[GR_DIM] = {0};
+                    double A_untilt_embed[GR_DIM] = {0}, A_tilt_embed[GR_DIM] = {0};
+                    G.coords.con_vec_to_embed(Xnative_midplane, A_untilt, A_untilt_embed);
+                    rotate_polar_vec(Xmidplane, A_untilt_embed, -tilt, Xembed, A_tilt_embed);
+                    G.coords.con_vec_to_native(Xnative, A_tilt_embed, A_tilt);
+
+                    // Lower the result as we need curl(A_mu).  Done at local zone.
+                    double A_tilt_lower[GR_DIM] = {0};
+                    G.lower(A_tilt, A_tilt_lower, k, j, i, Loci::corner);
+                    VLOOP A(v, k, j, i) = A_tilt_lower[1 + v];
+                } else {
+                    // Some problems rely on a very accurate A->B, which the rotation lacks.
+                    // So, we preserve exact values in the no-tilt case.
+                    A(V3, k, j, i) = Aphi;
+                }
+            });
+
+        if (pkgs.count("B_CT"))
+        {
+            auto B_Uf = rc->PackVariables(std::vector<std::string>{"cons.fB"});
+            // This fills a couple zones outside the exact interior with bad data
+            // Careful of that w/e.g. Dirichlet bounds.
+            IndexRange3 bB = KDomain::GetRange(rc, domain, 0, -1);
+            if (ndim > 2) {
+                pmb->par_for(
+                    "ot_B", bB.ks, bB.ke, bB.js, bB.je, bB.is, bB.ie,
+                    KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                        B_CT::curl_3D(G, A, B_Uf, k, j, i);
+                    });
+            } else if (ndim > 1) {
+                pmb->par_for(
+                    "ot_B", bB.ks, bB.ke, bB.js, bB.je, bB.is, bB.ie,
+                    KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                        B_CT::curl_2D(G, A, B_Uf, k, j, i);
+                    });
+            } else {
+                throw std::runtime_error("Must initialize 1D field directly!");
+            }
+            B_CT::BlockUtoP(rc, domain);
+        } else if (pkgs.count("B_FluxCT")) {
+            // Calculate B-field
+            GridVector B_U = rc->Get("cons.B").data;
+            IndexRange3 bl = KDomain::GetRange(rc, domain, 0, -1); // TODO will need changes if domain < entire
+            if (ndim > 2) {
+                pmb->par_for(
+                    "B_field_B_3D", bl.ks, bl.ke, bl.js, bl.je, bl.is, bl.ie,
+                    KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                        B_FluxCT::averaged_curl_3D(G, A, B_U, k, j, i);
+                    });
+            } else if (ndim > 1) {
+                pmb->par_for(
+                    "B_field_B_2D", bl.ks, bl.ke, bl.js, bl.je, bl.is, bl.ie,
+                    KOKKOS_LAMBDA(const int &k, const int &j, const int &i) {
+                        B_FluxCT::averaged_curl_2D(G, A, B_U, k, j, i);
+                    });
+            } else {
+                throw std::runtime_error("Must initialize 1D field directly!");
+            }
+            // Finally, make sure we initialize the primitive field too
+            B_FluxCT::BlockUtoP(rc, domain);
+        }
+
+        return TaskStatus::complete;
+    }
+}
+
 TaskStatus SeedBField(MeshData<Real> *md, ParameterInput *pin)
 {
     Flag("SeedBField");
@@ -81,7 +345,7 @@ TaskStatus SeedBField(MeshData<Real> *md, ParameterInput *pin)
     auto pmesh = md->GetMeshPointer();
     const int verbose = pmesh->packages.Get("Globals")->Param<int>("verbose");
 
-    if (verbose) {
+    if (MPIRank0() && verbose) {
         std::cout << "Seeding B field with type " << b_field_type << std::endl;
     }
 
@@ -115,6 +379,12 @@ TaskStatus SeedBField(MeshData<Real> *md, ParameterInput *pin)
             status = SeedBFieldType<BSeedType::bz_monopole>(rc, pin);
         } else if (b_field_type == "vertical") {
             status = SeedBFieldType<BSeedType::vertical>(rc, pin);
+        } else if (b_field_type == "orszag_tang") {
+            status = SeedBFieldType<BSeedType::orszag_tang>(rc, pin);
+        } else if (b_field_type == "orszag_tang_a") {
+            status = SeedBFieldType<BSeedType::orszag_tang_a>(rc, pin);
+        } else if (b_field_type == "wave") {
+            status = SeedBFieldType<BSeedType::wave>(rc, pin);
         } else {
             throw std::invalid_argument("Magnetic field seed type not supported: " + b_field_type);
         }
